// src/app/ngx-bit/lib/interface.ts
export interface NgxBitModuleOptions {
  origin: string;
  language?: any;
  withCredentials?: boolean;
  i18n?: any[];
  i18n_switch?: any[];
  page_limit?: number;
}

export interface I18nControlsOptions {
  value?: any;
  validate?: any;
  asyncValidate?: any;
}

export interface ListsFactoryOptions {
  index?: number;
  limit?: number;
  refresh?: boolean;
}